% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Correlation (Statistics)}
\usage{
correlation(rast1 = NULL, rast2 = NULL, size = 3,
  measure = "correlation", output = NULL, load_output = TRUE,
  quiet = TRUE, ...)
}
\arguments{
\item{rast1}{first \linkS4class{RasterLayer}}

\item{rast2}{second \linkS4class{RasterLayer}}

\item{measure}{Transformation type. Must be "correlation" or "covariance". Default: "correlation"}

\item{output}{(optional) file name to save output. Default: NULL}

\item{load_output}{If TRUE, function output file will be loaded into R. Default: TRUE}

\item{quiet}{IF FALSE, console messages are shown. Default: TRUE}

\item{...}{optional parameter for [RQGIS::run_qgis()]}
}
\value{
\linkS4class{RasterLayer} correlation
}
\description{
Calculates corrlation or covariance between two rasters.
}
\keyword{correlation}
