% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissection.R
\name{dissection}
\alias{dissection}
\title{Dissection (Surface Texture)}
\usage{
dissection(elevation = NULL, size = 3, output = NULL,
  load_output = TRUE, quiet = TRUE, ...)
}
\arguments{
\item{elevation}{\linkS4class{RasterLayer} containing elevation values}

\item{output}{(optional) file name to save output. Default: NULL}

\item{load_output}{If TRUE, function output file will be loaded into R. Default: TRUE}

\item{quiet}{IF FALSE, console messages are shown. Default: TRUE}

\item{...}{optional parameter for [RQGIS::run_qgis()]}
}
\value{
\linkS4class{RasterLayer} dissection
}
\description{
[dissection()] describes dissection in a continuous raster surface
within rectangular or circular window. Martonne’s modified dissection is calculated
as: d = ( z - z(min)) / (z(max) - z(min)).
}
\keyword{dissection}
