% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landform.R
\name{landform}
\alias{landform}
\title{Landform (Surface Texture)}
\usage{
landform(elevation = NULL, size = 3, output = NULL, load_output = TRUE,
  quiet = TRUE, ...)
}
\arguments{
\item{elevation}{\linkS4class{RasterLayer} containing elevation values}

\item{size}{analysis window (rectangle). Default: 3}

\item{output}{(optional) file name to save output. Default: NULL}

\item{load_output}{If TRUE, function output file will be loaded into R. Default: TRUE}

\item{quiet}{IF FALSE, console messages are shown. Default: TRUE}

\item{...}{optional parameter for [RQGIS::run_qgis()]}
}
\value{
\linkS4class{RasterLayer} landform
}
\description{
Concavity/convexity landform index (Bolstad’s variant).
}
\keyword{landform}
