% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAspect.R
\name{linearAspect}
\alias{linearAspect}
\title{Linear Aspect (Directionality)}
\usage{
linearAspect(elevation = NULL, aspect = NULL, size = 3, zscale = "1.0",
  output = NULL, load_output = TRUE, quiet = TRUE, ...)
}
\arguments{
\item{elevation}{\linkS4class{RasterLayer} containing elevation values}

\item{aspect}{(optional) \linkS4class{RasterLayer} containing aspect values. Note: North must be 0, and East must be 90. Default: NULL}

\item{size}{analysis window (rectangle). Default: 3}

\item{zscale}{Multiplicative factor to convert elevation units to horizontal units. Default: "1.0"}

\item{output}{(optional) file name to save output. Default: NULL}

\item{load_output}{If TRUE, function output file will be loaded into R. Default: TRUE}

\item{quiet}{IF FALSE, console messages are shown. Default: TRUE}

\item{...}{optional parameter for [RQGIS::run_qgis()]}
}
\value{
\linkS4class{RasterLayer} linear aspect
}
\description{
Transforms the circular aspect to a linear variable.
}
\keyword{aspect}
\keyword{linear}
