% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughness.R
\name{roughness}
\alias{roughness}
\title{Roughness (Surface Texture)}
\usage{
roughness(elevation = NULL, size = 3, output = NULL, load_output = TRUE,
  quiet = TRUE, ...)
}
\arguments{
\item{elevation}{\linkS4class{RasterLayer} containing elevation values}

\item{size}{analysis window (rectangle). Default: 3}

\item{output}{(optional) file name to save output. Default: NULL}

\item{load_output}{If TRUE, function output file will be loaded into R. Default: TRUE}

\item{quiet}{IF FALSE, console messages are shown. Default: TRUE}

\item{...}{optional parameter for [RQGIS::run_qgis()]}
}
\value{
\linkS4class{RasterLayer} roughness
}
\description{
[roughness()] represents the variance in a continuous raster within a specified window and scale.
}
\keyword{roughness}
