% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughnessIndex.R
\name{roughnessIndex}
\alias{roughnessIndex}
\title{Roughness Index (Surface Texture)}
\usage{
roughnessIndex(elevation = NULL, size = 3, output = NULL,
  load_output = TRUE, return_weighted = FALSE, quiet = TRUE, ...)
}
\arguments{
\item{elevation}{\linkS4class{RasterLayer} containing elevation values}

\item{output}{(optional) file name to save output. If return_weighted is TRUE, output should be vector containing two file names. Default: NULL}

\item{load_output}{If TRUE, function output file will be loaded into R. Default: TRUE}

\item{return_weighted}{If TRUE roughness index is normalized by max-value. Default: FALSE}

\item{quiet}{IF FALSE, console messages are shown. Default: TRUE}

\item{...}{optional parameter for [RQGIS::run_qgis()]}
}
\value{
\linkS4class{RasterLayer} roughness index. If return_weighted is TRUE, a list is returned containing roughness index [[1]] and normalized roughness index [[2]]
}
\description{
[roughnessIndex()] as standard deviation of residual topography (Cavalli et al. 2008).
}
\keyword{mean}
\keyword{slope}
