% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saTransformation.R
\name{saTransformation}
\alias{saTransformation}
\title{Slope/Aspect Transformation (Temperature and Moisture)}
\usage{
saTransformation(elevation = NULL, slope = NULL, aspect = NULL,
  zscale = "1.0", transType = "COS", output = NULL, load_output = TRUE,
  quiet = TRUE, ...)
}
\arguments{
\item{elevation}{\linkS4class{RasterLayer} containing elevation values}

\item{slope}{(optional) \linkS4class{RasterLayer} containing slope values. Value format must be percent. Default: NULL}

\item{aspect}{(optional) \linkS4class{RasterLayer} containing aspect values. Note: North must be 0, and East must be 90. Default: NULL}

\item{zscale}{Multiplicative factor to convert elevation units to horizontal units. Default: "1.0"}

\item{transType}{Transformation type. Must be "COS", "SIN" or "TRASP". Default: "COS"}

\item{output}{(optional) file name to save output. Default: NULL}

\item{load_output}{If TRUE, function output file will be loaded into R. Default: TRUE}

\item{quiet}{IF FALSE, console messages are shown. Default: TRUE}

\item{...}{optional parameter for [RQGIS::run_qgis()]}
}
\value{
\linkS4class{RasterLayer} saTransformation
}
\description{
Options are Stage’s (1976) COS, SIN; or Roberts & Cooper (1989) TRASP (topographic radiation aspect index).
COS AND SIN - An a priori assumption of a maximum in the NW quadrant (45 azimuth) and a minimum in the SW quadrant can be
replaced by an empirically determined location of the optimum (Stage, 1976). For slopes from 0% - 100%, the functions are
linearized and bounded from -1 to 1. Greater than 100% slopes are treated out of the -1 to 1 range and the model sets all
values greater than 100% to 101% and flat areas (-1) to nodata. TRASP - Circular aspect is transformed to assign a value of
zero to land oriented in a north- northeast direction, (typically the coolest and wettest orientation), and a value of one on
the hotter, dryer south-southwesterly slopes. The result is a continuous variable between 0 - 1 (Roberts and Cooper 1989).
}
\keyword{slope/aspect}
\keyword{transformation}
