% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopePosition.R
\name{slopePosition}
\alias{slopePosition}
\title{Slope Position (Surface Texture)}
\usage{
slopePosition(elevation = NULL, size = 3, output = NULL,
  load_output = TRUE, quiet = TRUE, ...)
}
\arguments{
\item{elevation}{\linkS4class{RasterLayer} containing elevation values}

\item{size}{analysis window (rectangle). Default: 3}

\item{output}{(optional) file name to save output. Default: NULL}

\item{load_output}{If TRUE, function output file will be loaded into R. Default: TRUE}

\item{quiet}{IF FALSE, console messages are shown. Default: TRUE}

\item{...}{optional parameter for [RQGIS::run_qgis()]}
}
\value{
\linkS4class{RasterLayer} slopePosition
}
\description{
[slopePosition()] calculates scalable slope position by subtracting a focalmean raster
from the original elevation raster.
}
\keyword{position}
\keyword{slope}
